New AI is complete!
In addition to previous logic, this new enemy is now capable of fleeing
when it reaches low health. When it runs far away enough from the
player, it will begin healing in pulses. Once its health goes to the
maximum, it will revert back to aggressive phase where it begins using
its 3 attacks against the player (fire, ice which immobilizes the
player, and it can summon enemies).