Modifications to Skeleton's Artificial Intelligence, bug fixes
Skeleton will now have a new ability which increases navmesh agent speed
and attack power. Shares a cooldown with melee attack. Can be used only
once it leaves idle state, but does not need to be in range of the
player.
Bug fix - in wizard artificial intelligence script, fixed an error when
it attempts to instantiate an ice shield as a child object in its
hierarchy. This change ensures the shield will go into the correct
wizard's gameObject hierarchy